+++ src/index.ts
@@ -1,7 +1,12 @@
 import * as babelTypes from "@babel/types";
 import Obfuscator from "./obfuscator";
 import { ObfuscateOptions } from "./options";
-import { ObfuscationResult, ProfilerCallback } from "./obfuscationResult";
+import {
+  ObfuscationResult,
+  ProfileData,
+  ProfilerCallback,
+  ProfilerLog,
+} from "./obfuscationResult";
 
 export async function obfuscate(
   sourceCode: string,
@@ -28,17 +33,51 @@ export async function obfuscateWithProfiler(
     callback: ProfilerCallback;
     performance: { now(): number };
   }
-): Promise<ObfuscationResult> {
+): Promise<ObfuscationResult & { profileData: ProfileData }> {
+  const startTime = performance.now();
+
   var obfuscator = new Obfuscator(options);
+  var totalTransforms = obfuscator.plugins.length;
 
-  var ast = obfuscator.parseCode(sourceCode);
+  var transformTimeMap: { [transformName: string]: number } =
+    Object.create(null);
+  var currentTransformTime = performance.now();
 
-  ast = obfuscator.obfuscateAST(ast, profiler.callback);
+  const beforeParseTime = performance.now();
 
-  var code = obfuscator.generateCode(ast);
+  var ast = Obfuscator.parseCode(sourceCode);
+
+  const parseTime = performance.now() - beforeParseTime;
+
+  ast = obfuscator.obfuscateAST(ast, {
+    profiler: (log: ProfilerLog) => {
+      var nowTime = performance.now();
+      transformTimeMap[log.currentTransform] = nowTime - currentTransformTime;
+      currentTransformTime = nowTime;
+      profiler.callback(log);
+    },
+  });
+
+  const beforeCompileTime = performance.now();
+
+  var code = Obfuscator.generateCode(ast, obfuscator.options);
+
+  const compileTime = performance.now() - beforeCompileTime;
+
+  const endTime = performance.now();
+
+  const obfuscationTime = endTime - startTime;
 
   return {
     code: code,
+    profileData: {
+      transformTimeMap: transformTimeMap,
+      obfuscationTime: obfuscationTime,
+      parseTime: parseTime,
+      compileTime: compileTime,
+      totalTransforms: totalTransforms,
+      totalPossibleTransforms: obfuscator.totalPossibleTransforms,
+    },
   };
 }